+++ src/CVMapMicro.cpp
@@ -1,6 +1,7 @@
 #include "plugin.hpp"
 #include "MapModuleBase.hpp"
 #include "components/MapButton.hpp"
+#include "components/VoltageLedDisplay.hpp"
 #include <chrono>
 
 namespace StoermelderPackOne {
@@ -102,6 +103,10 @@ struct CVMapMicroModule : CVMapModuleBase<1> {
 		CVMapModuleBase<1>::process(args);
 	}
 
+	float getCurrentVoltage() {
+		return inputs[INPUT].getVoltage();
+	}
+
 	json_t* dataToJson() override {
 		json_t* rootJ = CVMapModuleBase<1>::dataToJson();
 		json_object_set_new(rootJ, "panelTheme", json_integer(panelTheme));
@@ -126,10 +131,15 @@ struct CVMapMicroWidget : ThemedModuleWidget<CVMapMicroModule> {
 		addChild(createWidget<StoermelderBlackScrew>(Vec(RACK_GRID_WIDTH, 0)));
 		addChild(createWidget<StoermelderBlackScrew>(Vec(RACK_GRID_WIDTH, RACK_GRID_HEIGHT - RACK_GRID_WIDTH)));
 
-		MapButton<CVMapMicroModule>* button = createParamCentered<MapButton<CVMapMicroModule>>(Vec(22.5f, 60.3f), module, CVMapMicroModule::MAP_PARAM);
+		VoltageLedDisplay<CVMapMicroModule>* ledDisplay = createWidgetCentered<VoltageLedDisplay<CVMapMicroModule>>(Vec(22.5f, 43.0f));
+		ledDisplay->box.size = Vec(39.1f, 13.2f);
+		ledDisplay->module = module;
+		addChild(ledDisplay);
+
+		MapButton<CVMapMicroModule>* button = createParamCentered<MapButton<CVMapMicroModule>>(Vec(22.5f, 78.8f), module, CVMapMicroModule::MAP_PARAM);
 		button->setModule(module);
 		addParam(button);
-		addChild(createLightCentered<MapLight<GreenRedLight>>(Vec(22.5f, 60.3f), module, CVMapMicroModule::MAP_LIGHT));
+		addChild(createLightCentered<MapLight<GreenRedLight>>(Vec(22.5f, 78.8f), module, CVMapMicroModule::MAP_LIGHT));
 
 		addInput(createInputCentered<StoermelderPort>(Vec(22.5f, 142.0f), module, CVMapMicroModule::INPUT));